Update madison_wcb.py
Corrected small typo on line 146 (representing)

I'm also wondering if students could be allowed to input WCB size -- up to 500 x 360 -- to accommodate other sizes of paper (I have the normal 11x8.5, 10x7, 6.875x5, and 5.5x5.5). They could eyeball it, but it would be easier to make sure their painting would fit on the paper if the screen matches.